Overall Increment in background color
This changes make the color theme look a bit natural and are more reliable to human eyes. Matt Colors are scientifically tested to be more appealing to Eyes. You can imagine if the keys of your keyboard are shiny and are of a direct black color they would look weird. A little matt finish makes them more original